daily_bot/requirements.txt
Django==4.1.3
psycopg2-binary==2.9.3
SQLAlchemy==1.4.39
databases[postgresql]==0.4.2
python-decouple==3.3
django-cors-headers==3.13.0
slackclient==2.9.4
djangorestframework==3.14.0
celery==5.2.7
Redis==4.4.0
slack-sdk==3.19.5
django-celery-beat==2.4.0
